linux_initial_requirements.txt
certifi==2021.5.30
click==8.0.4
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.5
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
pandas==1.1.5
Pillow @ file:///tmp/build/80754af9/pillow_1625670622947/work
python-dateutil==2.9.0.post0
pytz==2024.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
